backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
stripe==7.6.0
tensorflow==2.15.0
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-whois==0.8.0
dnspython==2.4.2 